Fixed story edge which wanted int vertices not long vertices

+++ src/Phantonia.Historia/StoryEdge.cs
@@ -1,6 +1,6 @@
 ﻿namespace Phantonia.Historia;
 
-public readonly struct StoryEdge(int toVertex, int fromVertex, bool isWeak)
+public readonly struct StoryEdge(long toVertex, long fromVertex, bool isWeak)
 {
     public long ToVertex { get; } = toVertex;
